+++ src/ska_integration_test_harness/facades/tmc_central_node_facade.py
@@ -4,7 +4,6 @@
 from typing import Tuple
 
 import tango
-from deprecated import deprecated
 from ska_control_model import ResultCode
 
 from ska_integration_test_harness.actions.central_node.central_node_assign_resources import (  # pylint: disable=line-too-long # noqa E501
@@ -28,20 +27,14 @@ from ska_integration_test_harness.actions.central_node.move_to_on import (
 from ska_integration_test_harness.actions.central_node.set_standby import (
     SetStandby,
 )
+from ska_integration_test_harness.common_utils.deprecated import DeprecatedMeta
 from ska_integration_test_harness.inputs.json_input import JSONInput
 from ska_integration_test_harness.structure.telescope_wrapper import (
     TelescopeWrapper,
 )
 
 
-@deprecated(
-    reason=(
-        "This facade is deprecated, since we want to use a facade "
-        "for each subsystem. Please use TMCFacade instead."
-    ),
-    version="0.2.0",
-)
-class TMCCentralNodeFacade:
+class TMCCentralNodeFacade(metaclass=DeprecatedMeta):
     """A facade to TMC Central Node device and its actions.
 
     A facade to TMC subsystem, providing a simplified interface to the
@@ -60,13 +53,6 @@ class TMCCentralNodeFacade:
     instead.
     """
 
-    @deprecated(
-        reason=(
-            "This facade is deprecated, since we want to use a facade "
-            "for each subsystem. Please use TMCFacade instead."
-        ),
-        version="0.2.0",
-    )
     def __init__(self, telescope: TelescopeWrapper) -> None:
         self._telescope = telescope