+++ src/ska_integration_test_harness/facades/tmc_subarray_node_facade.py
@@ -1,7 +1,6 @@
 # pylint: disable=duplicate-code
 """A facade to subarray in TMC."""
 
-from deprecated import deprecated
 from ska_control_model import ObsState
 
 from ska_integration_test_harness.actions.subarray.force_change_of_obs_state import (  # pylint: disable=line-too-long # noqa: E501
@@ -40,6 +39,7 @@ from ska_integration_test_harness.actions.subarray.subarray_restart import (  # 
 from ska_integration_test_harness.actions.subarray.subarray_scan import (  # pylint: disable=line-too-long # noqa: E501
     SubarrayScan,
 )
+from ska_integration_test_harness.common_utils.deprecated import DeprecatedMeta
 from ska_integration_test_harness.inputs.json_input import JSONInput
 from ska_integration_test_harness.inputs.test_harness_inputs import (
     TestHarnessInputs,
@@ -49,14 +49,7 @@ from ska_integration_test_harness.structure.telescope_wrapper import (  # pylint
 )
 
 
-@deprecated(
-    reason=(
-        "This facade is deprecated, since we want to use a facade "
-        "for each subsystem. Please use TMCFacade instead."
-    ),
-    version="0.2.0",
-)
-class TMCSubarrayNodeFacade:
+class TMCSubarrayNodeFacade(metaclass=DeprecatedMeta):
     """A facade to TMC Subarray Node device and its actions.
 
     A facade to TMC subsystem, providing a simplified interface to the
@@ -77,13 +70,6 @@ class TMCSubarrayNodeFacade:
     instead.
     """
 
-    @deprecated(
-        reason=(
-            "This facade is deprecated, since we want to use a facade "
-            "for each subsystem. Please use TMCFacade instead."
-        ),
-        version="0.2.0",
-    )
     def __init__(self, telescope: TelescopeWrapper) -> None:
         self._telescope = telescope