+++ src/ska_integration_test_harness/structure/telescope_wrapper.py
@@ -1,9 +1,5 @@
 """A wrapper class that contains all the telescope subsystems."""
 
-from assertpy import assert_that
-from ska_control_model import ObsState
-from ska_tango_testing.integration import TangoEventTracer
-
 from ska_integration_test_harness.common_utils.tango_devices_info import (
     DevicesInfoProvider,
     DevicesInfoServiceException,
@@ -71,17 +67,6 @@ class TelescopeWrapper:
 
     _instance = None
 
-    _tmc: TMCWrapper | None = None
-    _sdp: SDPWrapper | None = None
-    _csp: CSPWrapper | None = None
-    _dishes: DishesWrapper | None = None
-
-    devices_info_provider: DevicesInfoProvider | None = None
-    """The devices info provider used to access the devices information.
-
-    (Used for recap purposes).
-    """
-
     def __new__(cls):
         if cls._instance is None:
             cls._instance = super(TelescopeWrapper, cls).__new__(cls)
@@ -89,6 +74,11 @@ class TelescopeWrapper:
 
     # -----------------------------------------------------------------
     # Subsystem access points
+
+    _tmc: TMCWrapper | None = None
+    _sdp: SDPWrapper | None = None
+    _csp: CSPWrapper | None = None
+    _dishes: DishesWrapper | None = None
 
     @property
     def tmc(self) -> TMCWrapper:
@@ -130,6 +120,15 @@ class TelescopeWrapper:
         self.fail_if_not_set_up()
         return self._dishes
 
+    # -----------------------------------------------------------------
+    # Recap generation tools
+
+    devices_info_provider: DevicesInfoProvider | None = None
+    """The devices info provider used to access the devices information.
+
+    (Used for recap purposes).
+    """
+
     def get_subsystems_recap(self, update_devices_info: bool = True) -> str:
         """Get a recap of the active subsystems and their devices.
 
@@ -165,6 +164,9 @@ class TelescopeWrapper:
     # -----------------------------------------------------------------
     # Initialisation and tear down methods
 
+    actions_default_timeout: int = 60
+    """The default timeout (in seconds) used in the telescope actions."""
+
     def set_up(
         self,
         tmc: TMCWrapper,
@@ -183,10 +185,6 @@ class TelescopeWrapper:
         self._sdp = sdp
         self._csp = csp
         self._dishes = dishes
-        # TODO: Add here some "health checks" (?)
-
-    TEARDOWN_TIMEOUT = 50
-    """The timeout (in seconds) used in the teardown procedure."""
 
     def tear_down(self) -> None:
         """Tear down the entire telescope test structure.
@@ -195,43 +193,10 @@ class TelescopeWrapper:
         """
         self.fail_if_not_set_up()
 
-        # Subscribe to the obsState events to be sure that
-        # teardown moves the system to the expected state.
-        event_tracer = TangoEventTracer()
-        event_tracer.subscribe_event(self.tmc.subarray_node, "obsState")
-        event_tracer.subscribe_event(self.csp.csp_subarray, "obsState")
-        event_tracer.subscribe_event(self.sdp.sdp_subarray, "obsState")
-
         self.tmc.tear_down()
         self.sdp.tear_down()
         self.csp.tear_down()
         self.dishes.tear_down()
-
-        # Assert the system is in the expected state after the teardown
-        assert_that(event_tracer).described_as(
-            "FAIL IN TEARDOWN PROCEDURE: "
-            "TMC Subarray node "
-            f"({self.tmc.subarray_node}) "
-            "ObsState is supposed to be EMPTY after the teardown."
-        ).within_timeout(self.TEARDOWN_TIMEOUT).has_change_event_occurred(
-            self.tmc.subarray_node, "obsState", ObsState.EMPTY
-        )
-        assert_that(event_tracer).described_as(
-            "FAIL IN TEARDOWN PROCEDURE: "
-            "CSP Subarray node "
-            f"({self.csp.csp_subarray}) "
-            "obsState is supposed to be EMPTY after the teardown."
-        ).within_timeout(self.TEARDOWN_TIMEOUT).has_change_event_occurred(
-            self.csp.csp_subarray, "obsState", ObsState.EMPTY
-        )
-        assert_that(event_tracer).described_as(
-            "FAIL IN TEARDOWN PROCEDURE: "
-            "SDP Subarray node "
-            f"({self.sdp.sdp_subarray}) "
-            "obsState is supposed to be EMPTY after the teardown."
-        ).within_timeout(self.TEARDOWN_TIMEOUT).has_change_event_occurred(
-            self.sdp.sdp_subarray, "obsState", ObsState.EMPTY
-        )
 
     def fail_if_not_set_up(self) -> None:
         """Fail if the telescope test structure is not set up.